+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.10.0] - 2020-03-09
 ### Added
 - `fullWidth` to `SliderLayout` schema definition and allow it to be controlled by the Site Editor.
 - `autoplay` and `itemsPerPage` to `SliderLayout` schema as `layout` properties.